add tabs lib
