+++ package.json
@@ -12,7 +12,6 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.25.1",
-    "react-tabs": "^6.1.1",
     "react-transition-group": "^4.4.5"
   },
   "devDependencies": {
