+++ src/App.tsx
@@ -2,14 +2,8 @@ import 'bulma/css/bulma.css';
 import '@fortawesome/fontawesome-free/css/all.css';
 import './App.scss';
 import { Link, Navigate, Route, Routes, useLocation } from 'react-router-dom';
-import HomePage from './components/HomePage/HomePage';
-import { TabsPage } from './components/TabsPage/TabsPage';
-
-const tabs = [
-  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
-  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
-  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
-];
+import HomePage from './pages/HomePage';
+import { TabsPage } from './pages/TabsPage';
 
 export const App = () => {
   const { pathname } = useLocation();
@@ -44,8 +38,8 @@ export const App = () => {
             <Route path="/" element={<HomePage />} />
             <Route path="/home" element={<Navigate to="/" replace />} />
             <Route path="tabs">
-              <Route index element={<TabsPage tabs={tabs} />} />
-              <Route path=":tabId" element={<TabsPage tabs={tabs} />} />
+              <Route index element={<TabsPage />} />
+              <Route path=":tabId" element={<TabsPage />} />
             </Route>
             <Route
               path="*"
