--- src/components/TabsPage/TabsPage.tsx
@@ -1,37 +1,0 @@
-import { Link, useParams } from 'react-router-dom';
-import classNames from 'classnames';
-import { Tab } from '../../types/Tab';
-
-type Props = {
-  tabs: Tab[];
-};
-
-export const TabsPage = ({ tabs }: Props) => {
-  const { tabId } = useParams();
-
-  const currentTab = tabs.find(tab => tab.id === tabId);
-
-  return (
-    <>
-      <h1 className="title">Tabs page</h1>
-
-      <div className="tabs is-boxed">
-        <ul>
-          {tabs.map(tab => (
-            <li
-              key={tab.id}
-              data-cy="Tab"
-              className={classNames({ 'is-active': tab.id === tabId })}
-            >
-              <Link to={`/tabs/${tab.id}`}>{tab.title}</Link>
-            </li>
-          ))}
-        </ul>
-      </div>
-
-      <div className="block" data-cy="TabContent">
-        {currentTab ? currentTab.content : 'Please select a tab'}
-      </div>
-    </>
-  );
-};
