+++ src/pages/TabsPage.tsx
@@ -1,0 +1,20 @@
+import { useParams } from 'react-router-dom';
+import { Tab } from '../types/Tab';
+import { Tabs } from '../components/Tabs';
+
+const tabs: Tab[] = [
+  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
+  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
+  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
+];
+
+export const TabsPage = () => {
+  const { tabId } = useParams();
+
+  return (
+    <>
+      <h1 className="title">Tabs page</h1>
+      <Tabs tabs={tabs} selectedTabId={tabId} />
+    </>
+  );
+};
